togyzkumalak-engine/requirements.txt
# Togyzkumalak Engine Requirements

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# ML/AI
torch>=2.0.0
numpy<2.0

# Gym environment
gym>=0.26.0

# Google Gemini (optional)
google-generativeai>=0.3.0

# Utilities
pydantic>=2.0.0
python-multipart>=0.0.6

